backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
redis
httpx
requests
python-multipart
